fix: ensures login trace for sesame links
